pyproject.toml
[tool.poetry]
name = "vtw-ai-assistant-and-ocr"
version = "0.1.0"
description = "vtw ai assistant"
authors = ["closeup1202 <closeup1202@gmail.com>"]
license = "none"
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.4"
python-dotenv = "^1.0.1"
streamlit = "^1.38.0"
torch = {version = "^2.4.0+cu124", source = "pytorch-gpu-src"}
torchaudio = {version = "^2.4.0+cu124", source = "pytorch-gpu-src"}
easyocr = "^1.7.1"
langchain-upstage = "^0.1.8"
langchain-community = "^0.2.14"
torchvision = "^0.19.0"
langchain-pinecone = "^0.1.3"
langchain-text-splitters = "^0.2.2"
python-docx = "^1.1.2"
langchain-ollama = "^0.1.3"
langchain-huggingface = "^0.0.3"
pyperclip = "^1.9.0"
st-copy-to-clipboard = "^0.1.6"


[[tool.poetry.source]]
name = "pytorch-gpu-src"
url = "https://download.pytorch.org/whl/cu124"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
